+++ src/index.ts
@@ -2,3 +2,4 @@ export * from './health.constants';
 export * from './health.interface';
 export * from './health.controller';
 export * from './health.module';
+export * from './metrics';