+++ src/metrics/index.ts
@@ -1,0 +1,5 @@
+export * from './metrics.constants';
+export * from './metrics.interface';
+export * from './metrics.module';
+export * from './metrics.controller';
+export * from './metrics.service';