+++ src/metrics/metrics.constants.ts
@@ -1,0 +1,1 @@
+export const CUSTOM_METRICS = 'CUSTOM_METRICS';